LPS-201086 Merge AuditMessageFactoryUtil/AuditMessageFactoryImpl into AuditMessageFactory
# breaking_change_report

## What
portal-kernel/src/com/liferay/portal/kernel/audit/AuditMessageFactory.java
portal-kernel/src/com/liferay/portal/kernel/audit/AuditMessageFactoryUtil.java

AuditMessageFactoryUtil and AuditMessageFactoryImpl logic are merged into AuditMessageFactory

## Why

This interface/impl/util separation does not have any value. Merge it into single class to avoid the unnecessary module.
