--- modules/.releng/apps/portal-security-audit/portal-security-audit-impl/artifact.properties
@@ -1,3 +1,0 @@
-artifact.git.id=31496e606478b3bc78698330f5194c7bb9b75107
-artifact.sources.url=https://repository-cdn.liferay.com/nexus/content/repositories/liferay-public-releases/com/liferay/com.liferay.portal.security.audit.impl/4.0.8/com.liferay.portal.security.audit.impl-4.0.8-sources.jar
-artifact.url=https://repository-cdn.liferay.com/nexus/content/repositories/liferay-public-releases/com/liferay/com.liferay.portal.security.audit.impl/4.0.8/com.liferay.portal.security.audit.impl-4.0.8.jar
