--- modules/.releng/apps/portal-security-audit/portal-security-audit-impl/liferay-releng.changelog
@@ -1,34 +1,0 @@
-#
-# Bundle Version 3.0.0
-#
-d7b046eaf8e6dd3b689d5ce0590f24efa3be5ac1^..13bb3beec1507dcad9cced1937329b81c16d69d5=LPS-77425 LPS-80332
-
-#
-# Bundle Version 3.0.6
-#
-13bb3beec1507dcad9cced1937329b81c16d69d5^..0928cfd80e30de284d8c7a2dd068d3a329cadcd6=LPS-105380
-
-#
-# Bundle Version 4.0.0
-#
-0928cfd80e30de284d8c7a2dd068d3a329cadcd6^..c86bd0af84b96588f4cfc7761432438a302e506b=LPS-121824
-
-#
-# Bundle Version 4.0.4
-#
-c86bd0af84b96588f4cfc7761432438a302e506b^..ea6822857db63c5b0a528aea69a836ccc0eeb144=LPS-105380
-
-#
-# Bundle Version 4.0.6
-#
-ea6822857db63c5b0a528aea69a836ccc0eeb144^..6d1efe1c96758c8cf79300d68d3bc003b1ed9c19=LPS-168433
-
-#
-# Bundle Version 4.0.7
-#
-6d1efe1c96758c8cf79300d68d3bc003b1ed9c19^..abcdbc514b2f5d708bdd4335b581795545fc4489=LPS-180466
-
-#
-# Bundle Version 4.0.8
-#
-abcdbc514b2f5d708bdd4335b581795545fc4489^..ba111f6976812ce2cd535c93a13bd82c8b45e297=LPS-188912
