--- modules/apps/portal-security-audit/portal-security-audit-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal Security Audit Implementation
-Bundle-SymbolicName: com.liferay.portal.security.audit.impl
-Bundle-Version: 4.0.9
