+++ modules/modules.properties
@@ -1687,7 +1687,6 @@ bundle.version[com.liferay.portal.security.audit.api]=8.0.2
 bundle.version[com.liferay.portal.security.audit.event.generators.api]=6.2.3
 bundle.version[com.liferay.portal.security.audit.event.generators.user.management]=5.0.13
 bundle.version[com.liferay.portal.security.audit.event.generators]=5.0.17
-bundle.version[com.liferay.portal.security.audit.impl]=4.0.9
 bundle.version[com.liferay.portal.security.audit.router]=6.0.21
 bundle.version[com.liferay.portal.security.audit.storage.api]=8.0.2
 bundle.version[com.liferay.portal.security.audit.storage.service]=6.0.40
@@ -3679,7 +3678,6 @@ project.name[com.liferay.portal.security.audit.api]=:apps:portal-security-audit:
 project.name[com.liferay.portal.security.audit.event.generators.api]=:apps:portal-security-audit:portal-security-audit-event-generators-api
 project.name[com.liferay.portal.security.audit.event.generators.user.management]=:apps:portal-security-audit:portal-security-audit-event-generators-user-management
 project.name[com.liferay.portal.security.audit.event.generators]=:dxp:apps:portal-security-audit:portal-security-audit-event-generators
-project.name[com.liferay.portal.security.audit.impl]=:apps:portal-security-audit:portal-security-audit-impl
 project.name[com.liferay.portal.security.audit.router]=:apps:portal-security-audit:portal-security-audit-router
 project.name[com.liferay.portal.security.audit.storage.api]=:apps:portal-security-audit:portal-security-audit-storage-api
 project.name[com.liferay.portal.security.audit.storage.service]=:apps:portal-security-audit:portal-security-audit-storage-service
