+++ portal-kernel/src/com/liferay/portal/kernel/audit/AuditMessageFactory.java
@@ -13,34 +13,63 @@ import java.util.Date;
 /**
  * @author Amos Fong
  */
-public interface AuditMessageFactory {
+public class AuditMessageFactory {
 
-	public AuditMessage getAuditMessage(String message) throws JSONException;
+	public AuditMessage getAuditMessage(String message) throws JSONException {
+		return new AuditMessage(message);
+	}
 
 	public AuditMessage getAuditMessage(
 		String eventType, long companyId, long groupId, long userId,
 		String userName, String className, String classPK, String message,
-		Date timestamp, JSONObject additionalInfoJSONObject);
+		Date timestamp, JSONObject additionalInfoJSONObject) {
+
+		return new AuditMessage(
+			eventType, companyId, groupId, userId, userName, className, classPK,
+			message, timestamp, additionalInfoJSONObject);
+	}
 
 	public AuditMessage getAuditMessage(
-		String eventType, long companyId, long userId, String userName);
+		String eventType, long companyId, long userId, String userName) {
+
+		return new AuditMessage(eventType, companyId, userId, userName);
+	}
 
 	public AuditMessage getAuditMessage(
 		String eventType, long companyId, long userId, String userName,
-		String className, String classPK);
+		String className, String classPK) {
+
+		return new AuditMessage(
+			eventType, companyId, userId, userName, className, classPK);
+	}
 
 	public AuditMessage getAuditMessage(
 		String eventType, long companyId, long userId, String userName,
-		String className, String classPK, String message);
+		String className, String classPK, String message) {
+
+		return new AuditMessage(
+			eventType, companyId, userId, userName, className, classPK,
+			message);
+	}
 
 	public AuditMessage getAuditMessage(
 		String eventType, long companyId, long userId, String userName,
 		String className, String classPK, String message, Date timestamp,
-		JSONObject additionalInfoJSONObject);
+		JSONObject additionalInfoJSONObject) {
+
+		return new AuditMessage(
+			eventType, companyId, userId, userName, className, classPK, message,
+			timestamp, additionalInfoJSONObject);
+	}
 
 	public AuditMessage getAuditMessage(
 		String eventType, long companyId, long userId, String userName,
 		String className, String classPK, String message,
-		JSONObject additionalInfoJSONObject);
+		JSONObject additionalInfoJSONObject) {
+
+		return new AuditMessage(
+			eventType, companyId, userId, userName, className, classPK, message,
+			additionalInfoJSONObject);
+	}
 
 }
